Feature/integrate linting pair program

See merge request sygnum/blockchain-engineering/ethereum/solidity-edge-contracts/solidity-edge-dchf-contracts!26

+++ contracts/Migrations.sol
@@ -1,9 +1,8 @@
 pragma solidity 0.5.12;
 
-
 contract Migrations {
     address public owner;
-    uint public lastCompletedMigration;
+    uint256 public lastCompletedMigration;
 
     modifier restricted() {
         if (msg.sender == owner) {
@@ -15,7 +14,7 @@ contract Migrations {
         owner = msg.sender;
     }
 
-    function setCompleted(uint completed) public restricted {
+    function setCompleted(uint256 completed) public restricted {
         lastCompletedMigration = completed;
     }
